fix andes icon provider signature
